refactor: Move code from query-popup
All code related feature information is now moved to feature-info widget.
