+++ projects/hslayers/src/components/query/query-popup-feature/feature-widgets/feature-info.component.html
