+++ projects/hslayers/src/components/query/query-popup-feature/feature-widgets/feature-info.component.ts
@@ -1,17 +1,17 @@
 import {ChangeDetectionStrategy, Component, Input, OnInit} from '@angular/core';
 
-import {HsConfirmDialogComponent} from '../../../../../common/confirm/confirm-dialog.component';
-import {HsDialogContainerService} from '../../../../layout/dialogs/dialog-container.service';
-import {HsLanguageService} from '../../../../language/language.service';
-import {HsQueryPopupServiceModel} from '../../../query-popup.service.model';
-import {HsQueryPopupWidgetBaseComponent} from '../../../query-popup-widget-base.component';
-import {HsQueryVectorService} from '../../../query-vector.service';
+import {HsConfirmDialogComponent} from '../../../../common/confirm/confirm-dialog.component';
+import {HsDialogContainerService} from '../../../layout/dialogs/dialog-container.service';
+import {HsLanguageService} from '../../../language/language.service';
+import {HsQueryPopupServiceModel} from '../../query-popup.service.model';
+import {HsQueryPopupWidgetBaseComponent} from '../../query-popup-widget-base.component';
+import {HsQueryVectorService} from '../../query-vector.service';
 import {
   getFeatureLabel,
   getFeatureName,
   getFeatureTitle,
   getFeatures,
-} from '../../../../../common/feature-extensions';
+} from '../../../../common/feature-extensions';
 
 @Component({
   selector: 'hs-feature-info',
