+++ projects/hslayers/src/components/query/query-popup-feature/query-popup-feature.component.html
@@ -1,0 +1,4 @@
+<form>
+  <hs-panel-container [service]="hsFeatureWidgetContainerService" [ownerComponent]="this">
+  </hs-panel-container>
+</form>
