+++ projects/hslayers/src/components/query/query-popup/query-popup.component.html
@@ -1,33 +1,3 @@
-<!-- Feature header template -->
-<ng-template #featureHeader let-feature="feature" let-type="type">
-        <div class="d-flex flex-row">
-            <div class="p-1 flex-grow-1"><em>{{serializeFeatureName(feature)}}</em></div>
-            <a class="p-1" *ngIf="HsQueryVectorService.isFeatureRemovable(feature) && type"
-                (click)="removeFeature(feature)" data-toggle="tooltip"
-                [title]="'QUERY.featurePopup.' + type | translate">
-                <i class="icon-remove-circle" style="color: rgb(228, 99, 99)"></i>
-            </a>
-        </div>
-</ng-template>
-
-<!-- Feature attribute row template -->
-<ng-template #featureAttribute let-attributes="attributes">
-        <div *ngFor="let attribute of attributes" class="d-flex flex-row">
-            <ng-container *ngIf="attribute.key">
-                <div class="p-1 flex-grow-1">
-                    {{HsLanguageService.getTranslationIgnoreNonExisting(
-                    'FEATURE_TABLE',
-                    attribute.key
-                    )}} 
-                </div>
-            </ng-container>
-            <ng-container *ngIf="attribute.displayFunction">
-                <div class="p-1" [innerHtml]="attribute.displayFunction(attribute.value)"></div>
-            </ng-container>
-        </div>
-</ng-template>
-
-
 <div [ngStyle]="popupVisible()" class="hs-hover-popup">
     <div class="card" style="padding: 5px;">
         <div *ngFor="let layerDesc of data.service.featureLayersUnderMouse; let i = index"
@@ -37,38 +7,9 @@
                 <i class="icon-remove-circle" style="color: rgb(73, 80, 87)"></i>
             </a>
             <strong style="font-size: small;">{{HsLayerUtilsService.translateTitle(layerDesc.title)}}</strong>
-            <hs-feature-popup-layer [layerDescriptor]="layerDesc" [service]="data.service"></hs-feature-popup-layer>
+            <hs-query-popup-layer [layerDescriptor]="layerDesc" [service]="data.service"></hs-query-popup-layer>
             <ng-container *ngFor="let feature of layerDesc.features">
-                <!-- Layout for clustered layers -->
-                <div *ngIf="isClustered(feature); else nonClusteredPopUp"
-                    style="font-size: small; line-height: 1; margin-bottom: 2px; border-bottom: 2px solid rgb(199, 199, 199)">
-                    <div class="d-flex flex-column">
-                        <ng-container *ngTemplateOutlet="featureHeader;context:{feature: feature}"></ng-container>
-                        <div [hidden]="!(attributesForHover.length > 0)" class="ms-1 mt-1">
-                            <ng-container *ngFor="let attribute of attributesForHover">
-                                <div *ngIf="attribute.feature"
-                                    [ngClass]="{'d-flex flex-column hs-hover-popup-feature' : hasMultipleSubFeatures(feature)}">
-                                    <ng-container
-                                        *ngTemplateOutlet="featureHeader;context:{feature: attribute.feature, type: 'removeClusterPart'}">
-                                    </ng-container>
-                                    <ng-container
-                                        *ngTemplateOutlet="featureAttribute;context:{attributes: attribute.attributes}">
-                                    </ng-container>
-                                </div>
-                            </ng-container>
-                        </div>
-                    </div>
-                </div>
-                <!-- Layout for non-clustered layers -->
-                <ng-template #nonClusteredPopUp>
-                    <div class="hs-hover-popup-feature">
-                        <ng-container
-                            *ngTemplateOutlet="featureHeader;context:{feature: feature, type: 'removeFeature'}">
-                        </ng-container>
-                        <ng-container *ngTemplateOutlet="featureAttribute;context:{attributes: HsQueryBaseService.serializeFeatureAttributes(feature)}">
-                        </ng-container>
-                    </div>
-                </ng-template>
+                <hs-query-popup-feature [feature]="feature" [attributesForHover]="attributesForHover" [service]="data.service"></hs-query-popup-feature>
             </ng-container>
         </div>
     </div>
