+++ projects/hslayers/src/components/query/query-popup/query-popup.component.ts
@@ -8,23 +8,15 @@ import {
 
 import {Subscription} from 'rxjs';
 
-import {HsConfirmDialogComponent} from '../../../common/confirm/confirm-dialog.component';
 import {HsDialogComponent} from '../../layout/dialogs/dialog-component.interface';
-import {HsDialogContainerService} from '../../layout/dialogs/dialog-container.service';
+
 import {HsDialogItem} from '../../layout/dialogs/dialog-item';
 import {HsEventBusService} from '../../core/event-bus.service';
 import {HsLanguageService} from '../../language/language.service';
 import {HsLayerUtilsService} from '../../utils/layer-utils.service';
 import {HsMapService} from '../../map/map.service';
-import {HsQueryBaseService} from '../query-base.service';
 import {HsQueryPopupServiceModel} from '../query-popup.service.model';
-import {HsQueryVectorService} from '../query-vector.service';
-import {
-  getFeatureLabel,
-  getFeatureName,
-  getFeatureTitle,
-  getFeatures,
-} from '../../../common/feature-extensions';
+import {getFeatures} from '../../../common/feature-extensions';
 import {getPopUp} from '../../../common/layer-extensions';
 
 @Component({
@@ -32,7 +24,8 @@ import {getPopUp} from '../../../common/layer-extensions';
   templateUrl: './query-popup.component.html',
 })
 export class HsQueryPopupComponent
-  implements OnDestroy, HsDialogComponent, AfterViewInit {
+  implements OnDestroy, HsDialogComponent, AfterViewInit
+{
   getFeatures = getFeatures;
   olMapLoadsSubscription: Subscription;
   attributesForHover = [];
@@ -43,12 +36,9 @@ export class HsQueryPopupComponent
   };
 
   constructor(
-    public HsQueryBaseService: HsQueryBaseService,
-    public HsQueryVectorService: HsQueryVectorService,
     public HsEventBusService: HsEventBusService,
     public HsLanguageService: HsLanguageService,
     public HsLayerUtilsService: HsLayerUtilsService, //Used in template
-    public HsDialogContainerService: HsDialogContainerService,
     public HsMapService: HsMapService,
     private ElementRef: ElementRef
   ) {
@@ -58,9 +48,11 @@ export class HsQueryPopupComponent
       }
     );
   }
+
   ngAfterViewInit(): void {
     this.data.service.registerPopup(this.ElementRef.nativeElement);
   }
+
   ngOnDestroy(): void {
     this.HsMapService.map.removeOverlay(this.data.service.hoverPopup);
     this.olMapLoadsSubscription.unsubscribe();
@@ -82,14 +74,14 @@ export class HsQueryPopupComponent
       let tmpForHover: any[] = [];
       this.data.service.featuresUnderMouse.forEach((feature) => {
         tmpForHover = tmpForHover.concat(
-          this.HsQueryBaseService.serializeFeatureAttributes(feature)
+          this.data.service.serializeFeatureAttributes(feature)
         );
         if (getFeatures(feature)) {
           getFeatures(feature).forEach((subfeature) => {
             const subFeatureObj: any = {};
             subFeatureObj.feature = subfeature;
             subFeatureObj.attributes =
-              this.HsQueryBaseService.serializeFeatureAttributes(subfeature);
+              this.data.service.serializeFeatureAttributes(subfeature);
             tmpForHover.push(subFeatureObj);
           });
         }
@@ -100,49 +92,5 @@ export class HsQueryPopupComponent
     return {
       'display': featureCount > 0 ? 'block' : 'none',
     };
-  }
-
-  isClustered(feature): boolean {
-    return getFeatures(feature) && getFeatures(feature).length > 0;
-  }
-
-  hasMultipleSubFeatures(feature): boolean {
-    return getFeatures(feature).length > 1;
-  }
-
-  serializeFeatureName(feature): string {
-    if (!feature) {
-      return;
-    }
-    if (getFeatureName(feature)) {
-      return getFeatureName(feature);
-    }
-    if (getFeatureTitle(feature)) {
-      return getFeatureTitle(feature);
-    }
-    if (getFeatureLabel(feature)) {
-      return getFeatureLabel(feature);
-    }
-    if (getFeatures(feature)) {
-      return this.HsLanguageService.getTranslation('QUERY.clusterContaining', {
-        count: getFeatures(feature).length,
-      });
-    }
-    return this.HsLanguageService.getTranslation('QUERY.untitledFeature');
-  }
-
-  async removeFeature(feature): Promise<void> {
-    const dialog = this.HsDialogContainerService.create(
-      HsConfirmDialogComponent,
-      {
-        message: this.HsLanguageService.getTranslation('QUERY.reallyDelete'),
-        title: this.HsLanguageService.getTranslation('QUERY.confirmDelete'),
-      }
-    );
-    const confirmed = await dialog.waitResult();
-    if (confirmed == 'yes') {
-      this.HsQueryVectorService.removeFeature(feature);
-      this.data.service.featuresUnderMouse = [];
-    }
   }
 }
