+++ projects/hslayers/src/components/query/query.module.ts
@@ -7,7 +7,7 @@ import {TranslateModule} from '@ngx-translate/core';
 
 import {HsClearLayerComponent} from './query-popup-layer/layer-widgets/clear-layer.component';
 import {HsDownloadModule} from '../../common/download/download.module';
-import {HsFeatureInfoComponent} from './query-popup-feature/feature-widgets/feature-info/feature-info.component';
+import {HsFeatureInfoComponent} from './query-popup-feature/feature-widgets/feature-info.component';
 import {HsPanelHelpersModule} from '../layout/panels/panel-helpers.module';
 import {HsQueryAttributeRowComponent} from './attribute-row/attribute-row.component';
 import {HsQueryComponent} from './query.component';
